refactor infer

+++ crates/builtin/dataset/src/labeled.rs
@@ -3,5 +3,5 @@ use vm::AnyValueDyn;
 
 pub struct LabeledData<'eval> {
     pub input: Arc<dyn AnyValueDyn<'eval>>,
-    pub label: usize,
+    pub label: i32,
 }
